refactor(ui): refactor utility cards
